add AdminController
