+++ templates/admin/index.html.twig
@@ -8,7 +8,7 @@
 
 {% block body %}
     <main class="container">
-        <form action="{{ path('app_song_index') }}" method="post">
+        <form action="{{ path('admin_index') }}" method="post">
             <div class="form-row">
                 <div class="form-group d-flex justify-content-end col-9 col-md-7 col-lg-5 col-xl-3 m-auto">
                     <input type="text" name="title" class="search-input form-control mb-2" placeholder="rechercher un titre ...">
@@ -22,12 +22,12 @@
         </form>
 
         <div class="d-flex flex-column flex-md-row align-items-center justify-content-center">
-            <form method="post" action="{{ path('app_song_index') }}">
+            <form method="post" action="{{ path('admin_index') }}">
                 <input type="hidden" name="isApproved" value="false" />
                 <button type="submit" class="button-unapproved btn">En attente de validation</button>
             </form>
 
-            <form method="post" action="{{ path('app_song_index') }}">
+            <form method="post" action="{{ path('admin_index') }}">
                 <input type="hidden" name="isApproved" value="true" />
                 <button type="submit" class="button-approved btn">Validation effectué</button>
             </form>
